Add type annotations to StatMech.jl

+++ src/StatMech.jl
@@ -20,21 +20,21 @@ export bose_einstein_distribution,
     subsystem_entropy,
     subsystem_volumetric_specific_heat
 
-const HBAR = 1
-const BOLTZMANN = 2
+const HBAR = 1.0
+const BOLTZMANN = 2.0
 
-validate_frequency(frequency) = frequency < 0 && throw(DomainError("Negative frequency is not proper for QHA calculation!"))
+validate_frequency(frequency::Float64) = frequency < 0 && throw(DomainError("Negative frequency is not proper for QHA calculation!"))
 
-bose_einstein_distribution(temperature, frequency) = 1 / (exp(HBAR * frequency / (BOLTZMANN * temperature)) - 1)
+bose_einstein_distribution(temperature::Float64, frequency::Float64) = 1 / (exp(HBAR * frequency / (BOLTZMANN * temperature)) - 1)
 
-function subsystem_partition_function(temperature, frequency)
+function subsystem_partition_function(temperature::Float64, frequency::Float64)
     frequency == 0 && return 1
 
     x = HBAR * frequency / (BOLTZMANN * temperature)
     return exp(x / 2) / (exp(x) - 1)
 end
 
-function subsystem_free_energy(temperature, frequency)
+function subsystem_free_energy(temperature::Float64, frequency::Float64)
     validate_frequency(frequency)
     frequency == 0 && return 0
 
@@ -43,7 +43,7 @@ function subsystem_free_energy(temperature, frequency)
     return hw / 2 + kt * log(1 - exp(-hw / kt))
 end
 
-function subsystem_internal_energy(temperature, frequency)
+function subsystem_internal_energy(temperature::Float64, frequency::Float64)
     validate_frequency(frequency)
     frequency == 0 && return BOLTZMANN * temperature
 
@@ -51,14 +51,14 @@ function subsystem_internal_energy(temperature, frequency)
     return hw / 2 / tanh(hw / (2 * BOLTZMANN * temperature))
 end
 
-function subsystem_entropy(temperature, frequency)
+function subsystem_entropy(temperature::Float64, frequency::Float64)
     validate_frequency(frequency)
 
     n = bose_einstein_distribution(temperature, frequency)
     return BOLTZMANN * ((1 + n) * log(1 + n) - n * log(n))
 end
 
-function subsystem_volumetric_specific_heat(temperature, frequency)
+function subsystem_volumetric_specific_heat(temperature::Float64, frequency::Float64)
     validate_frequency(frequency)
     frequency == 0 && return BOLTZMANN
